requirements.txt
disnake==2.9.1
python-dotenv==1.0.0
coloredlogs==15.0.1
asyncio==3.4.3
aiomysql==0.2.0
PyNaCl==1.5.0